UiaTextRange::GetText bugfix where incorrect row text was being returned

Cherry picked from Pull request 555672.

Related work items: #12329833

+++ src/interactivity/win32/ut_interactivity_win32/UiaTextRangeTests.cpp
@@ -282,7 +282,7 @@ class UiaTextRangeTests
 
         for (auto it = rowMappings.begin(); it != rowMappings.end(); ++it)
         {
-            VERIFY_ARE_EQUAL(it->second, _range->_normalizeRow(it->first));
+            VERIFY_ARE_EQUAL(static_cast<int>(it->second), _range->_normalizeRow(it->first));
         }
     }
